requirements.txt
pandas>=2.0
numpy>=1.26
pydantic>=2.10
python-Levenshtein>=0.21.1
pytest>=8.0
streamlit>=1.28.0
openpyxl>=3.1.0
xlrd>=2.0.1
PyPDF2>=3.0.1
tabulate>=0.9.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
